split out diff toolbar to another page
